Support CS API 0.4.0
Numerous changes in CS (and now also AS) API, including compatibility-breaking ones - see the diff for details.

+++ lib/csapi/content-repo.cpp
@@ -284,7 +284,7 @@ BaseJob::Status GetUrlPreviewJob::parseJson(const QJsonDocument& data)
 class GetConfigJob::Private
 {
     public:
-        Omittable<double> uploadSize;
+        Omittable<qint64> uploadSize;
 };
 
 QUrl GetConfigJob::makeRequestUrl(QUrl baseUrl)
@@ -304,7 +304,7 @@ GetConfigJob::GetConfigJob()
 
 GetConfigJob::~GetConfigJob() = default;
 
-Omittable<double> GetConfigJob::uploadSize() const
+Omittable<qint64> GetConfigJob::uploadSize() const
 {
     return d->uploadSize;
 }
@@ -312,7 +312,7 @@ Omittable<double> GetConfigJob::uploadSize() const
 BaseJob::Status GetConfigJob::parseJson(const QJsonDocument& data)
 {
     auto json = data.object();
-    d->uploadSize = fromJson<double>(json.value("m.upload.size"_ls));
+    d->uploadSize = fromJson<qint64>(json.value("m.upload.size"_ls));
     return Success;
 }
