+++ lib/csapi/content-repo.h
@@ -248,7 +248,7 @@ namespace QMatrixClient
             /// The maximum size an upload can be in bytes.
             /// Clients SHOULD use this as a guide when uploading content.
             /// If not listed or null, the size limit should be treated as unknown.
-            Omittable<double> uploadSize() const;
+            Omittable<qint64> uploadSize() const;
 
         protected:
             Status parseJson(const QJsonDocument& data) override;
